+++ src/actions/question.mjs
@@ -7,11 +7,11 @@ import db from '../../server/db/db.mjs';
 export function editQuestion(id, question, options) {
     const questionDao = new QuestionDao(db);
     try {
-        if(question) {
+        if(question && id.toString().match("^\\d+$")) {
             const nUpdated = questionDao.editQuestion(id, question, options || []);
             return({nUpdated});
         } else {
-            return({error: "No question text provided."});
+            return({error: "No question text provided or invalid ID."});
         }
     } catch(e) {
         return({error: "Internal server error"});
@@ -19,11 +19,15 @@ export function editQuestion(id, question, options) {
 }
 
 export function deleteQuestion(id) {
-    const questionDao = new QuestionDao(db);
-    try {
-        const nUpdated = questionDao.deleteQuestion(id);
-        return {nUpdated};
-    } catch(e) {
-        return {error: "Internal server error"};
-    }
+	if(id.toString().match("^\\d+$")) {
+    	const questionDao = new QuestionDao(db);
+    	try {
+        	const nUpdated = questionDao.deleteQuestion(id);
+        	return {nUpdated};
+    	} catch(e) {
+        	return {error: "Internal server error"};
+    	}
+	} else {
+		return { error: "Invalid ID." };
+	}
 }
